move some doc utilites out of FileSystemDocInterpretor
